refactor: 계층형에서 도메인형으로 구조 변경
