+++ src/main/java/com/ceos19/springeverytime/domain/chatroom/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
-package com.ceos19.springeverytime.repository;
+package com.ceos19.springeverytime.domain.chatroom.repository;
 
-import com.ceos19.springeverytime.domain.ChatRoom;
+import com.ceos19.springeverytime.domain.chatroom.domain.ChatRoom;
 import com.ceos19.springeverytime.domain.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
