+++ src/main/java/com/ceos19/springeverytime/domain/comment/domain/Comment.java
@@ -1,6 +1,7 @@
-package com.ceos19.springeverytime.domain;
+package com.ceos19.springeverytime.domain.comment.domain;
 
-import com.ceos19.springeverytime.domain.like.CommentLike;
+import com.ceos19.springeverytime.domain.BaseEntity;
+import com.ceos19.springeverytime.domain.like.domain.CommentLike;
 import com.ceos19.springeverytime.domain.post.domain.Post;
 import com.ceos19.springeverytime.domain.user.domain.User;
 import jakarta.persistence.*;
