+++ src/main/java/com/ceos19/springeverytime/domain/image/domain/Image.java
@@ -1,5 +1,6 @@
-package com.ceos19.springeverytime.domain;
+package com.ceos19.springeverytime.domain.image.domain;
 
+import com.ceos19.springeverytime.domain.BaseEntity;
 import com.ceos19.springeverytime.domain.post.domain.Post;
 import jakarta.persistence.*;
 import lombok.*;
